fix(submission): read reference sequence as i32 to fix int4 decode
next_reference_seq read the DB result as i64, but the SQL returns an
INTEGER (int4) column. sqlx fails to decode int4 as i64, so the read
silently fell back to 1, generating the same reference on every retry
and colliding with the existing SUB-{year}-00001 row.

+++ backend/src/repositories/submission.rs
@@ -162,7 +162,7 @@ impl SubmissionRepository {
                 "SELECT COALESCE(MAX(CAST(SUBSTRING(reference FROM 'SUB-[0-9]+-([0-9]+)') AS INTEGER)), 0) + 1 AS next_seq FROM submissions WHERE reporting_year = $1 AND reference LIKE 'SUB-%'",
                 vec![sea_orm::Value::Int(Some(reporting_year))],
             );
-            let next_seq: i64 = self
+            let next_seq: i32 = self
                 .db
                 .query_one(stmt)
                 .await
